fix: api 파일 내 type 분리
