+++ briefin/src/api/newsApi.ts
@@ -1,33 +1,7 @@
 import { apiClient, type ApiResponse } from './client';
+import type { NewsListItem, NewsDetailResponse, ScrapResponse, NewsItem } from '@/types/news';
 
-export interface NewsListItem {
-  newsId: string;
-  title: string;
-  summary: string;
-  category: string;
-  region: string;
-  press: string;
-  publishedAt: string;
-  relatedCompanies: string[];
-}
-
-export interface NewsDetailResponse {
-  newsId: string;
-  title: string;
-  content: string;
-  summary: string;
-  category: string;
-  press: string;
-  publishedAt: string;
-  relatedCompanies: string[];
-  relatedNews: string[];
-}
-
-export interface ScrapResponse {
-  newsId: number;
-  isScraped: boolean;
-  scrapedAt: string;
-}
+export type { NewsListItem, NewsDetailResponse, ScrapResponse };
 
 // 뉴스 목록 조회 (category 없으면 전체)
 export const fetchNewsList = (category?: string) => {
@@ -60,8 +34,6 @@ export const deleteScrapNews = (id: string | number) =>
   apiClient.delete<ApiResponse<null>>(`/api/news/${id}/scrap`);
 
 // 백엔드 NewsListItem → 프론트 NewsItem 변환
-import type { NewsItem } from '@/types/news';
-
 export function toNewsItem(item: NewsListItem): NewsItem {
   return {
     id: String(item.newsId),
