+++ briefin/src/api/userApi.ts
@@ -1,26 +1,7 @@
 import { apiClient, type ApiResponse } from './client';
+import type { UserInfo, ScrapedNews, WatchlistCompany } from '@/types/mypage';
 
-export interface UserInfo {
-  userId: string;
-  email: string;
-  createdAt: string;
-}
-
-export interface ScrapedNews {
-  newsId: number;
-  title: string;
-  summary: string;
-  source: string;
-  scrapedAt: string;
-}
-
-export interface WatchlistCompany {
-  companyId: number;
-  companyName: string;
-  ticker: string;
-  logoUrl: string;
-  addedAt: string;
-}
+export type { UserInfo, ScrapedNews, WatchlistCompany };
 
 // 내 정보 조회
 export const fetchMyInfo = () =>
