+++ briefin/src/types/company.ts
@@ -1,3 +1,17 @@
+// API 응답 타입
+export interface CompanyDetail {
+  id: number;
+  name: string;
+  ticker: string;
+  sector: string;
+  logoUrl: string;
+  currentPrice: number;
+  changeRate: number;
+  marketCap: number;
+  isOverseas: boolean;
+  relatedCompanies: { id: number; name: string; ticker: string }[];
+}
+
 export type Company = {
   id: string;
   name: string;
